apq8084-common: Use the OSS GPS HAL
* Add cast in NAME_VAL definition to avoid gcc complaining

+++ gps/utils/loc_log.h
@@ -45,7 +45,7 @@ typedef struct
    long        val;
 } loc_name_val_s_type;
 
-#define NAME_VAL(x) {"" #x "", x }
+#define NAME_VAL(x) {"" #x "", (long)(x) }
 
 #define UNKNOWN_STR "UNKNOWN"
